% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Display map of fatalities}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer state number as defined by 
\href{https://www.census.gov/geo/reference/ansi_statetables.html}{the US Census Bureau}}

\item{year}{Integer specifying the year of interest}
}
\value{
A plot of all fatal accidents for the given state during the given year
}
\description{
Display map of fatalities
}
\note{
Using a state number that does not exist will throw an error. Also,
only the contiguous 48 states and District of Columbia can be mapped
}
\examples{
\dontrun{fars_map_state(26, 2015)}

}
