% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in a fars file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file containing fars data}
}
\value{
A data.frame containing fars data
}
\description{
Read in a fars file
}
\note{
The filename must exist or it will throw an error. Also, the file
must be able to be converted to a data.frame
}
\examples{
\dontrun{fars_read("accident_2015.csv.bz2")}
\dontrun{fars_read("c:/workspace/R/accident_2015.csv.bz2")}

}
