% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in one or more years of fars data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Vector of one or more years of interest}
}
\value{
List of data.frames, each row containing the months and year
}
\description{
Read in one or more years of fars data
}
\examples{
\dontrun{fars_read_years(2015)}
\dontrun{fars_read_years(c(2013, 2014, 2015))}

}
