% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the number of fatal accidents by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector of one or more years of interest}
}
\value{
a data.frame summarizing the number of fatal accidents by month and year
}
\description{
Summarize the number of fatal accidents by month and year
}
\examples{
\dontrun{fars_summarize_years(2015)}
\dontrun{fars_summarize_years(c(2013, 2014, 2015))}

}
