% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a fars filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Integer specifying the year of interest}
}
\value{
A properly formatted fars filename for the given year
}
\description{
Create a fars filename
}
\examples{
make_filename(2015)

}
